[Feature] 헤더, 좌측 사이드패널 Shell로 처리, 공고 조회 페이지 본문 크기 정리
